added a cleanup message to run through and clean up linked references
