+++ octgnFX/Octide/GameTab/MarkerTab/MarkerItemViewModel.cs
@@ -47,6 +47,11 @@ namespace Octide.ItemModel
             Asset.PropertyChanged += AssetUpdated;
             Name = m.Name;
         }
+        public override void Cleanup()
+        {
+            Asset.Cleanup();
+            base.Cleanup();
+        }
 
         private void AssetUpdated(object sender, PropertyChangedEventArgs e)
         {
@@ -56,12 +61,6 @@ namespace Octide.ItemModel
                 RaisePropertyChanged("Asset");
                 RaisePropertyChanged("Icon");
             }
-        }
-
-        public override void Cleanup()
-        {
-            Asset.SelectedAsset = null;
-            base.Cleanup();
         }
 
         public override object Clone()
