+++ octgnFX/Octide/PreviewTab/BoardItemModel.cs
@@ -55,6 +55,11 @@ namespace Octide.ItemModel
             Asset.PropertyChanged += AssetUpdated;
             Name = b.Name;
         }
+        public override void Cleanup()
+        {
+            Asset.Cleanup();
+            base.Cleanup();
+        }
 
         private void AssetUpdated(object sender, PropertyChangedEventArgs e)
         {
@@ -64,11 +69,6 @@ namespace Octide.ItemModel
                 RaisePropertyChanged("Asset");
                 RaisePropertyChanged("Icon");
             }
-        }
-        public override void Cleanup()
-        {
-            Asset.SelectedAsset = null;
-            base.Cleanup();
         }
 
         public override object Clone()
