+++ octgnFX/Octide/PreviewTab/PhaseItemModel.cs
@@ -44,6 +44,11 @@ namespace Octide.ItemModel
             Asset.PropertyChanged += AssetUpdated;
             Name = p.Name;
         }
+        public override void Cleanup()
+        {
+            Asset.Cleanup();
+            base.Cleanup();
+        }
 
         private void AssetUpdated(object sender, PropertyChangedEventArgs e)
         {
@@ -54,12 +59,6 @@ namespace Octide.ItemModel
                 RaisePropertyChanged("Icon");
             }
         }
-        public override void Cleanup()
-        {
-            Asset.SelectedAsset = null;
-            base.Cleanup();
-        }
-
         public override object Clone()
         {
             return new PhaseItemModel(this, Source);
