+++ octgnFX/Octide/ProxyTab/Templates/Link.cs
@@ -268,10 +268,5 @@ namespace Octide.ProxyTab.ItemModel
             return new TextLinkPropertyModel(Source);
         }
 
-        public override void Cleanup()
-        {
-            base.Cleanup();
-            Messenger.Default.Unregister<CustomPropertyChangedMessage>(this, CustomPropertyChanged);
-        }
     }
 }
